moved back into own folder and got some autogenerated stuff
