+++ front/package.json
@@ -3,6 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@types/node": "^14.14.41",
     "apollo-boost": "^0.3.1",
     "apollo-link": "^1.2.13",
     "apollo-link-context": "^1.0.19",
@@ -13,14 +14,14 @@
     "leaflet": "^1.6.0",
     "lodash": "^4.17.19",
     "polyline-encoded": "^0.0.9",
-    "react": "^16.12.0",
-    "react-apollo": "^2.5.8",
+    "react": "^17.0.2",
+    "react-apollo": "^3.1.5",
     "react-apollo-hooks": "^0.4.5",
-    "react-dom": "^16.12.0",
+    "react-dom": "^17.0.2",
     "react-leaflet": "^2.6.1",
     "react-lineto": "^3.2.0",
     "react-redux": "^7.1.3",
-    "react-scripts": "^2.1.8",
+    "react-scripts": "^4.0.3",
     "react-select": "^2.4.4",
     "react-timeline-semantic-ui": "^1.0.13",
     "redux": "^4.0.5",
@@ -43,7 +44,7 @@
     "overrides": [
       {
         "files": [
-          "**/*.stories.*"
+          "**/*.story.*"
         ],
         "rules": {
           "import/no-anonymous-default-export": "off"
@@ -66,9 +67,11 @@
     "@storybook/react": "^6.2.9",
     "@types/react": "^17.0.3",
     "@types/react-dom": "^17.0.3",
-    "gulp": "^4.0.2",
     "source-map-loader": "^2.0.1",
     "ts-loader": "^8.2.0",
     "typescript": "^4.2.4"
+  },
+  "resolutions": {
+    "@storybook/react/babel-loader": "8.1.0"
   }
 }
